Refactor main package name
