--- src/edu/psu/cmpsc/teamone/Instructor.java
@@ -1,5 +1,0 @@
-package edu.psu.cmpsc.teamone;
-
-public class Instructor {
-
-}
