--- src/edu/psu/cmpsc/teamone/Meeting.java
@@ -1,5 +1,0 @@
-package edu.psu.cmpsc.teamone;
-
-public class Meeting {
-
-}
