+++ src/edu/psu/teamone/main/Schedule.java
@@ -1,0 +1,5 @@
+package edu.psu.teamone.main;
+
+public class Schedule {
+
+}
